Fixed issue with summoned monsters not being properly initialised.
Added version LMPV2.0 to all new editor maps, enabling use of the existing Barony maps. All new maps created with the editor are now saved with LMPV2.0 in the header.

Moved item and equipment initialisation code back into the the initMonster functions instead of moving it elsewhere.

Added spare 32 editor flags for general use in the entity struct.

Added randomised stats settable in the editor, new function setRandomMonsterStats to be called after the stats are initialised.