Hopefully fix macOS crash on exit
The timerCallback function was trying to access players after it had
been freed by deinitGame. So let's stop the timer before freeing
players. This may still be susceptible to a race condition (timer is
removed successfully while timerCallback is still running, and
deinitGame manages to free it before timerCallback accesses it).
However, I think fixing all such corner cases is a bit of a lost cause
in this codebase… So here's something that should fix it *most* of the
time.