Remove music logic from map loading
This allows the editor to link successfully without having
sound_common.cpp (which would introduce a dependency on net.cpp).
This results in a little duplication — setting levelmusicplaying = false
on every in-game map load — which is not ideal, but still IMHO better
than making loadMap care so much about where it's called from. I hope to
improve this more later on.