Include short git hash in version string
This can make tracing issues back to the code a lot easier, especially
with steam versions, for people who do not know what happens to be the
current version on Steam etc., see #312

Also fix the version strings for the OSX bundle while we're at it. There
should only be one source of truth for the version number now.