Partially fixed level change desync in multiplayer.
The problem was that since the level number doesn't increase, the client doesn't know the level change packet it's receiving is legit or just the regular LVLC packet the server sends at intervals. So now the server is also sending whether or not it's on a secret level. So if currentlevel or secretlevel have changed, then the client loads the next map.
It works! Almost. The client loads in the next level, but it appears to keep their old X and Y instead of spawning them in the start room.