Added 11 new parameters for json modding
Added 11 new parameters for gameplaymodifiers json using existing functions in mod_tools.hpp.

The first 6 parameters allow modders to apply multipliers to each of the player's core stats in the various statGet functions.

The next two parameters allow modders to modify the base amount of HP and MP (current and max) the player gains upon leveling up.

The last three parameters introduce a new, optional mechanic for modders to play with: trauma damage. When enabled, a portion of damage taken is applied to the player's max HP, in addition to reducing their HP normally.

- doTraumaDamage is a boolean that determines whether this system is used at all. (default false)
- traumaDamagePercent determines the portion of damage dealt as trauma damage. (default 0%)
- traumaDamageHPLimit determines the minimum value trauma damage can reduce a player's max HP to. (default 25)